feat: daily temperature and precipitation
