+++ src/components/climate/Chart.js
@@ -1,8 +1,10 @@
 import Highcharts from 'highcharts'
-import highchartsMore from 'highcharts/highcharts-more'
+import exporting from 'highcharts/highcharts-more'
+import highchartsMore from 'highcharts/modules/exporting'
 import PropTypes from 'prop-types'
 import React, { useRef, useEffect } from 'react'
 
+exporting(Highcharts)
 highchartsMore(Highcharts)
 
 const Chart = ({ config }) => {
