+++ src/components/climate/ClimateChangeTab.js
@@ -1,14 +1,15 @@
 import PropTypes from 'prop-types'
 import React, { useRef, useEffect } from 'react'
+import Chart from './Chart.js'
 import DataLoading from './DataLoading.js'
 import ERA5Source from './ERA5Source.js'
-import TemperatureAnomaly from './TemperatureAnomaly.js'
+import getChart from './charts/temperatureAnomaly.js'
 
 // https://developers.google.com/earth-engine/datasets/catalog/NASA_GDDP-CMIP6
-const ClimateChange = ({ data }) => {
+const ClimateChangeTab = ({ data }) => {
     return data ? (
         <>
-            <TemperatureAnomaly data={data} />
+            <Chart config={getChart(data)} />
             <ERA5Source />
         </>
     ) : (
@@ -16,8 +17,8 @@ const ClimateChange = ({ data }) => {
     )
 }
 
-ClimateChange.propTypes = {
+ClimateChangeTab.propTypes = {
     data: PropTypes.array,
 }
 
-export default ClimateChange
+export default ClimateChangeTab
