+++ src/components/climate/styles/WeatherSymbol.module.css
@@ -4,6 +4,6 @@
 }
 
 .symbol img {
-    width: 40px;
-    height: 40px;
+    width: 38px;
+    height: 38px;
 }
